refactor(chat-citations): split citation UI into web and knowledge components
Extract WebCitation (panel card + hover content, incl. X oEmbed lazy
load) and KnowledgeCitation (panel card + URL-less hover content) into
components/chat/citations/, alongside shared citationPreview hooks and
common panel actions. CitationsPanel dispatches by citation type,
CitationsList slims down to the footer pill, and CitationTooltip
becomes a thin type dispatcher whose schema now tolerates URL-less
knowledge citations. Pure extraction — Tailwind classes and behaviour
unchanged; tests migrated alongside.
