+++ src/renderer/components/chat/messages/blocks/CitationsList.tsx
@@ -1,117 +1,18 @@
-import { Button, Scrollbar, Skeleton } from '@cherrystudio/ui'
+import { Button } from '@cherrystudio/ui'
 import Favicon from '@renderer/components/icons/FallbackFavicon'
-import SelectionContextMenu from '@renderer/components/SelectionContextMenu'
-import { useTemporaryValue } from '@renderer/hooks/useTemporaryValue'
-import { ipcApi } from '@renderer/ipc'
 import type { Citation } from '@renderer/types/message'
-import { fetchXOEmbed, isXPostUrl, xOembedKey } from '@renderer/utils/fetch'
-import { Check, Copy, FileSearch } from 'lucide-react'
+import { FileSearch } from 'lucide-react'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
-import { useSWRConfig } from 'swr'
-import useSWRImmutable from 'swr/immutable'
-import { v4 as uuid } from 'uuid'
 
+import { getCitationHostname } from '../../citations/common'
 import { useOptionalMessageListActions } from '../MessageListProvider'
-import type { MessageListActions } from '../types'
-
-type CitationCopyActions = Pick<MessageListActions, 'copyText' | 'notifyError'>
-type CitationPanelActions = CitationCopyActions & {
-  openPath?: (path: string) => void | Promise<void>
-  openExternalUrl?: (url: string) => void | Promise<void>
-}
 
 interface CitationsListProps {
   citations: Citation[]
 }
 
-interface CitationsPanelContentProps {
-  citations: Citation[]
-  actions?: CitationPanelActions
-}
-
-interface CitationPreviewSession {
-  load(url: string): Promise<void>
-}
-
-const citationPreviewKey = (url: string) => ['citationPreview', url] as const
-
-const useCitationPreviewSession = (): CitationPreviewSession => {
-  const { mutate } = useSWRConfig()
-  const [requestId] = React.useState<string>(() => uuid())
-  const requestsRef = React.useRef(new Map<string, Promise<void>>())
-
-  const load = React.useCallback(
-    (url: string): Promise<void> => {
-      const existing = requestsRef.current.get(url)
-      if (existing) return existing
-
-      const request = ipcApi
-        .request('citation.fetch_preview', { url, requestId })
-        .then(async ({ content }) => {
-          if (content) {
-            await mutate(citationPreviewKey(url), content, { revalidate: false })
-          }
-        })
-        .catch(() => undefined)
-
-      requestsRef.current.set(url, request)
-      return request
-    },
-    [mutate, requestId]
-  )
-
-  React.useEffect(() => {
-    const requests = requestsRef.current
-
-    return () => {
-      const hasRequests = requests.size > 0
-      requests.clear()
-
-      if (hasRequests) {
-        void ipcApi.request('citation.cancel_previews', { requestId }).catch(() => undefined)
-      }
-    }
-  }, [requestId])
-
-  return React.useMemo(() => ({ load }), [load])
-}
-
-const useCitationPreview = (url: string | undefined, session: CitationPreviewSession) => {
-  const { data } = useSWRImmutable<string>(url ? citationPreviewKey(url) : null, null)
-  const [settledUrl, setSettledUrl] = React.useState<string>()
-
-  React.useEffect(() => {
-    if (!url || data !== undefined) return
-
-    let active = true
-    void session.load(url).finally(() => {
-      if (active) setSettledUrl(url)
-    })
-
-    return () => {
-      active = false
-    }
-  }, [data, session, url])
-
-  return {
-    content: data,
-    isLoading: Boolean(url) && data === undefined && settledUrl !== url
-  }
-}
-
-const truncateText = (text: string, maxLength = 100) =>
-  text.length > maxLength ? `${text.slice(0, maxLength)}...` : text
-
-const getCitationHostname = (citation: Citation) => {
-  if (!citation.url) return undefined
-  try {
-    return new URL(citation.url).hostname
-  } catch {
-    return undefined
-  }
-}
-
+/** The "N sources" pill under a text block; opens the citations side panel. */
 const CitationsList: React.FC<CitationsListProps> = ({ citations }) => {
   const { t } = useTranslation()
   const openCitationsPanel = useOptionalMessageListActions()?.openCitationsPanel
@@ -149,189 +50,6 @@ const CitationsList: React.FC<CitationsListProps> = ({ citations }) => {
       <span className="h-3.5 w-px bg-border-subtle" />
       <span className="font-medium text-muted-foreground">{t('message.citation', { count })}</span>
     </Button>
-  )
-}
-
-export const CitationsPanelContent: React.FC<CitationsPanelContentProps> = ({ citations, actions }) => {
-  const previewSession = useCitationPreviewSession()
-
-  return (
-    <Scrollbar className="min-h-0 flex-1">
-      {citations.map((citation) => (
-        <div
-          key={`${citation.number}-${citation.url || citation.title}`}
-          className="border-border border-b-[0.5px] last:border-b-0">
-          {citation.type === 'websearch' && (
-            <div className="max-w-[min(400px,60vw)] px-3">
-              <WebSearchCitation citation={citation} previewSession={previewSession} actions={actions} />
-            </div>
-          )}
-          {citation.type === 'memory' && (
-            <div className="max-w-150 px-3">
-              <KnowledgeCitation citation={{ ...citation }} actions={actions} />
-            </div>
-          )}
-          {citation.type === 'knowledge' && (
-            <div className="max-w-150 px-3">
-              <KnowledgeCitation citation={{ ...citation }} actions={actions} />
-            </div>
-          )}
-        </div>
-      ))}
-    </Scrollbar>
-  )
-}
-
-const handleLinkClick = (
-  url: string,
-  event: React.MouseEvent,
-  actions?: {
-    openPath?: (path: string) => void | Promise<void>
-    openExternalUrl?: (url: string) => void | Promise<void>
-  }
-) => {
-  if (!url) return
-  if (url.startsWith('http')) {
-    if (!actions?.openExternalUrl) return
-    event.preventDefault()
-    void actions.openExternalUrl(url)
-    return
-  }
-
-  if (!actions?.openPath) return
-  event.preventDefault()
-  void actions.openPath(url)
-}
-
-const CopyButton: React.FC<{ content: string; actions?: CitationCopyActions }> = ({
-  content,
-  actions: injectedActions
-}) => {
-  const [copied, setCopied] = useTemporaryValue(false, 2000)
-  const { t } = useTranslation()
-  const actions = useOptionalMessageListActions()
-  const copyText = injectedActions?.copyText ?? actions?.copyText
-  const notifyError = injectedActions?.notifyError ?? actions?.notifyError
-
-  const handleCopy = () => {
-    if (!content || !copyText) return
-    Promise.resolve(copyText(content, { successMessage: t('common.copied') }))
-      .then(() => setCopied(true))
-      .catch(() => {
-        notifyError?.(t('message.copy.failed'))
-      })
-  }
-
-  if (!copyText) return null
-
-  return (
-    <div
-      className="-translate-y-1/2 absolute top-1/2 right-0 flex cursor-pointer items-center justify-center rounded p-1 text-muted-foreground opacity-0 transition-opacity duration-300 hover:bg-muted hover:opacity-100 group-hover:opacity-100"
-      onClick={handleCopy}>
-      {copied ? <Check size={14} /> : <Copy size={14} />}
-    </div>
-  )
-}
-
-const WebSearchCitation: React.FC<{
-  citation: Citation
-  previewSession: CitationPreviewSession
-  actions?: CitationPanelActions
-}> = ({ citation, previewSession, actions }) => {
-  const isXPost = Boolean(citation.url && isXPostUrl(citation.url))
-  const previewUrl = citation.url && !isXPost ? citation.url : undefined
-  const providerActions = useOptionalMessageListActions()
-  const linkActions = {
-    openPath: actions?.openPath ?? providerActions?.openPath,
-    openExternalUrl: actions?.openExternalUrl ?? providerActions?.openExternalUrl
-  }
-
-  const { content: previewContent, isLoading: isPreviewLoading } = useCitationPreview(previewUrl, previewSession)
-
-  const { data: oembedData, isLoading: isOembedLoading } = useSWRImmutable(
-    isXPost && citation.url ? xOembedKey(citation.url) : null,
-    () => fetchXOEmbed(citation.url),
-    { shouldRetryOnError: false }
-  )
-
-  const fetchedContent = isXPost
-    ? oembedData
-      ? truncateText(`@${oembedData.author}: ${oembedData.text}`)
-      : ''
-    : previewContent
-  const isLoading = isXPost ? isOembedLoading : isPreviewLoading
-  const displayTitle = isXPost && oembedData?.author ? `@${oembedData.author}` : citation.title
-  const titleContent = displayTitle || citation.hostname || citation.content || citation.url
-
-  return (
-    <SelectionContextMenu>
-      <div className="group relative flex w-full flex-col py-3 transition-all duration-300">
-        <div className="relative mb-1.5 flex w-full flex-row items-center gap-2">
-          {citation.showFavicon && getCitationHostname(citation) && (
-            <Favicon hostname={getCitationHostname(citation)!} alt={citation.title || citation.hostname || ''} />
-          )}
-          {citation.url ? (
-            <a
-              className="flex-1 text-nowrap text-foreground text-sm leading-[1.6] no-underline"
-              href={citation.url}
-              onClick={(e) => handleLinkClick(citation.url, e, linkActions)}>
-              {displayTitle || <span className="text-link">{citation.hostname}</span>}
-            </a>
-          ) : (
-            <span className="flex-1 text-nowrap text-foreground text-sm leading-[1.6]">{titleContent}</span>
-          )}
-
-          <div className="flex h-3.5 w-3.5 shrink-0 items-center justify-center rounded-full bg-primary/15 text-[10px] text-primary leading-[1.6] opacity-100 transition-opacity duration-300 group-hover:opacity-0">
-            {citation.number}
-          </div>
-          {fetchedContent && <CopyButton content={fetchedContent} actions={actions} />}
-        </div>
-        {isLoading ? (
-          <div className="space-y-1">
-            <Skeleton className="h-3 w-full" />
-            <Skeleton className="h-3 w-2/3" />
-          </div>
-        ) : (
-          fetchedContent && (
-            <div className="selectable-text cursor-text select-text break-all text-[13px] text-muted-foreground leading-[1.6]">
-              {fetchedContent}
-            </div>
-          )
-        )}
-      </div>
-    </SelectionContextMenu>
-  )
-}
-
-const KnowledgeCitation: React.FC<{ citation: Citation; actions?: CitationPanelActions }> = ({ citation, actions }) => {
-  const providerActions = useOptionalMessageListActions()
-  const linkActions = {
-    openPath: actions?.openPath ?? providerActions?.openPath,
-    openExternalUrl: actions?.openExternalUrl ?? providerActions?.openExternalUrl
-  }
-
-  return (
-    <SelectionContextMenu>
-      <div className="group relative flex w-full flex-col py-3 transition-all duration-300">
-        <div className="relative mb-1.5 flex w-full flex-row items-center gap-2">
-          {citation.showFavicon && <FileSearch width={16} />}
-          <a
-            className="flex-1 text-nowrap text-foreground text-sm leading-[1.6] no-underline"
-            href={citation.url}
-            onClick={(e) => handleLinkClick(citation.url, e, linkActions)}>
-            {/* example title: User/path/example.pdf */}
-            {citation.title?.split('/').pop()}
-          </a>
-          <div className="flex h-3.5 w-3.5 shrink-0 items-center justify-center rounded-full bg-primary/15 text-[10px] text-primary leading-[1.6] opacity-100 transition-opacity duration-300 group-hover:opacity-0">
-            {citation.number}
-          </div>
-          {citation.content && <CopyButton content={citation.content} actions={actions} />}
-        </div>
-        <div className="selectable-text cursor-text select-text break-all text-[13px] text-muted-foreground leading-[1.6]">
-          {citation.content ?? ''}
-        </div>
-      </div>
-    </SelectionContextMenu>
   )
 }
